* ⬆️ chore(dependabot.yml): add Dependabot configuration file * ⬆️ chore(Cargo.toml): upgrade package version to 0.1.1 and edition to 2023 * 🎨 style(rustfmt.toml): remove newline at end of file This commit adds a Dependabot configuration file to the repository, which specifies the package ecosystems to update and where the package manifests are located. The Cargo.toml file is also updated to version 0.1.1 and edition 2023. Finally, the rustfmt.toml file is updated to remove the newline at the end of the file.
* 🚀 feat(main.rs): add NavigationInfo resource
* 🐛 fix(main.rs): fix window title to include map id
The NavigationInfo resource is added to the App, which contains information about the available maps, the current map index, and the game path. This resource is used by the Navigation system to allow the player to switch between maps. The window title is updated to include the map id, which makes it easier for the player to identify which map they are currently on.

* 🐛 fix(main.rs): pass correct argument to get_map_ids function
The function get_map_ids was being passed an incorrect argument, which caused it to fail. The argument has been updated to the correct value, which resolves the issue.

* ✨ feat(systems): add navigation module
* 🐛 fix(navigation.rs): fix start_other_vakfu function argument type
The navigation module has been added to the systems module. It contains a NavigationInfo struct and a start_other_vakfu function. The NavigationInfo struct contains information about the current navigation state, including the map_ids, current_index, and game_path. The start_other_vakfu function takes a value and a game_path argument and starts another instance of the game with the specified map and game path. The value argument has been changed to an i32 type to match the expected type in the function.

* 🚀 feat(ui.rs): add navigation and exit functionality to ui_system
The ui_system function now takes in a NavigationInfo resource and an EventWriter for AppExit. This allows the function to navigate to other parts of the application and exit the application.

* ✨ feat(ui.rs): add navigation window to UI
The navigation window allows the user to navigate between maps and open a specific map. The window contains two buttons for navigating to the previous and next maps respectively, as well as a scrollable list of all available maps. When a map is clicked, the application will start with the selected map.